added toString() for account and account manager

+++ Split/src/main/java/split/ContactsManager.java
@@ -1,6 +1,4 @@
 package split;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
 import java.util.*;
 
 public class ContactsManager {
@@ -12,12 +10,16 @@ public class ContactsManager {
     }
 
     public void addEntry(UUID id) {
-        assertTrue(!contactsDatabase.containsKey(id));
-        contactsDatabase.put(id, new Contacts());
+        if (!contactsDatabase.containsKey(id)) {
+            contactsDatabase.put(id, new Contacts());
+        }
     }
 
     public void removeEntry(UUID id) {
-        assertTrue(contactsDatabase.containsKey(id));
-        contactsDatabase.remove(id);
+        if (contactsDatabase.containsKey(id)) {
+            contactsDatabase.remove(id);
+        }
     }
+
+    
 }
